big cleanup, leaving only working files
