--- integrated_report_template.html
@@ -1,149 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8">
-    <title>Integrated Analysis Report</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 20px;
-        }
-
-        h1 {
-            color: #333;
-        }
-
-        h2 {
-            color: #555;
-            border-bottom: 1px solid #ccc;
-            padding-bottom: 5px;
-        }
-
-        h3 {
-            color: #666;
-            margin-top: 30px;
-        }
-
-        h4 {
-            color: #777;
-            margin-top: 20px;
-        }
-
-        table {
-            border-collapse: collapse;
-            width: 100%;
-            margin-bottom: 20px;
-        }
-
-        th,
-        td {
-            text-align: right;
-            padding: 8px;
-            border: 1px solid #ddd;
-        }
-
-        th {
-            background-color: #f2f2f2;
-        }
-
-        img {
-            max-width: 100%;
-            height: auto;
-            margin-bottom: 20px;
-        }
-
-        .section {
-            margin-bottom: 60px;
-        }
-
-        .stat {
-            font-weight: bold;
-        }
-
-        p {
-            line-height: 1.6;
-        }
-    </style>
-</head>
-
-<body>
-    <h1>Integrated Analysis Report</h1>
-    <p><strong>Target Variable:</strong> {{ target_variable }}</p>
-
-    <!-- Numerical Variables Analysis -->
-    <div class="section">
-        <h2>Numerical Variables Analysis</h2>
-        <h3>Correlation Analysis</h3>
-        <h4>Correlation Matrix</h4>
-        <img src="{{ numerical_analysis.corr_heatmap }}" alt="Correlation Matrix Heatmap">
-        <h4>Correlation with Target Variable</h4>
-        <img src="{{ numerical_analysis.corr_barplot }}" alt="Correlation with Target Variable">
-        <h4>Correlation Coefficients</h4>
-        <table>
-            <tr>
-                <th>Variable</th>
-                <th>Correlation with {{ target_variable }}</th>
-            </tr>
-            {% for var, corr in numerical_analysis.target_corr.items() %}
-            <tr>
-                <td>{{ var }}</td>
-                <td>{{ corr | round(3) }}</td>
-            </tr>
-            {% endfor %}
-        </table>
-        <h3>Summary Statistics Grouped by {{ target_variable }}</h3>
-        {{ numerical_analysis.group_stats_html | safe }}
-        <h3>Boxplots of Top Variables by {{ target_variable }}</h3>
-        {% for plot in numerical_analysis.boxplots %}
-        <h4>{{ plot.variable }}</h4>
-        <img src="{{ plot.plot }}" alt="Boxplot of {{ plot.variable }} by {{ target_variable }}">
-        {% endfor %}
-    </div>
-
-    <!-- Categorical Variables Analysis -->
-    <div class="section">
-        <h2>Categorical Variables Analysis</h2>
-        {% for attr in categorical_analysis %}
-        <div class="subsection">
-            <h3>{{ attr.attribute }}</h3>
-            {% if attr.analysis.freq_counts_html %}
-            <h4>Frequency Counts</h4>
-            {{ attr.analysis.freq_counts_html | safe }}
-            {% endif %}
-            {% if attr.analysis.barchart %}
-            <h4>Bar Chart</h4>
-            <img src="{{ attr.analysis.barchart }}" alt="Bar Chart of {{ attr.attribute }}">
-            {% endif %}
-            {% if attr.relationship %}
-            <h4>Relationship with {{ target_variable }}</h4>
-            {% if attr.relationship.contingency_table_html %}
-            <h5>Contingency Table</h5>
-            {{ attr.relationship.contingency_table_html | safe }}
-            {% endif %}
-            {% if attr.relationship.plot %}
-            <h5>Stacked Bar Chart</h5>
-            <img src="{{ attr.relationship.plot }}"
-                alt="Stacked Bar Chart of {{ attr.attribute }} vs {{ target_variable }}">
-            {% endif %}
-            {% if attr.relationship.chi_squared %}
-            <h5>Chi-Squared Test</h5>
-            <p><strong>Chi-Squared Statistic:</strong> {{ attr.relationship.chi_squared.chi2_statistic | round(4) }}</p>
-            <p><strong>p-value:</strong> {{ attr.relationship.chi_squared.p_value | scientific_notation(4) }}</p>
-            <p><strong>Degrees of Freedom:</strong> {{ attr.relationship.chi_squared.degrees_of_freedom }}</p>
-            {% endif %}
-            {% if attr.relationship.cramers_v %}
-            <h5>Cramér's V</h5>
-            <p><strong>Cramér's V:</strong> {{ attr.relationship.cramers_v | round(3) }}</p>
-            <p><strong>Association Strength:</strong> {{ attr.relationship.association_strength }}</p>
-            {% endif %}
-            {% else %}
-            <p>No significant relationship with {{ target_variable }}.</p>
-            {% endif %}
-        </div>
-        {% endfor %}
-    </div>
-
-</body>
-
-</html>
