--- report_template.html
@@ -1,171 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8">
-    <title>Sequential Attribute Analysis Report</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 20px;
-        }
-
-        h1 {
-            color: #333;
-        }
-
-        h2 {
-            color: #555;
-            border-bottom: 1px solid #ccc;
-            padding-bottom: 5px;
-        }
-
-        h3 {
-            color: #666;
-            margin-top: 30px;
-        }
-
-        h4 {
-            color: #777;
-            margin-top: 20px;
-        }
-
-        table {
-            border-collapse: collapse;
-            width: 100%;
-            margin-bottom: 20px;
-        }
-
-        th,
-        td {
-            text-align: right;
-            padding: 8px;
-            border: 1px solid #ddd;
-        }
-
-        th {
-            background-color: #f2f2f2;
-        }
-
-        img {
-            max-width: 100%;
-            height: auto;
-            margin-bottom: 20px;
-        }
-
-        .section {
-            margin-bottom: 60px;
-        }
-
-        .stat {
-            font-weight: bold;
-        }
-
-        p {
-            line-height: 1.6;
-        }
-    </style>
-</head>
-
-<body>
-    <h1>Sequential Attribute Analysis Report</h1>
-    <p><strong>Target Variable:</strong> {{ target_variable }}</p>
-
-    {% for attr in attributes %}
-    <div class="section">
-        <h2>{{ loop.index }}. {{ attr.attribute }} ({{ attr.type.title() }})</h2>
-
-        {% if attr.type == 'numerical' %}
-        <!-- Descriptive Statistics -->
-        {% if attr.analysis.desc_stats %}
-        <h3>Descriptive Statistics</h3>
-        {{ attr.analysis.desc_stats.to_html(classes='table table-striped', header=False) | safe }}
-        {% endif %}
-        <!-- Histogram -->
-        {% if attr.analysis.histogram %}
-        <h3>Histogram</h3>
-        <img src="{{ attr.analysis.histogram }}" alt="Histogram of {{ attr.attribute }}">
-        {% endif %}
-        <!-- Boxplot -->
-        {% if attr.analysis.boxplot %}
-        <h3>Boxplot</h3>
-        <img src="{{ attr.analysis.boxplot }}" alt="Boxplot of {{ attr.attribute }}">
-        {% endif %}
-        {% elif attr.type == 'categorical' %}
-        <!-- Frequency Counts -->
-        {% if attr.analysis.freq_counts %}
-        <h3>Frequency Counts</h3>
-        {{ attr.analysis.freq_counts.to_html(classes='table table-striped', header=False) | safe }}
-        {% endif %}
-        <!-- Bar Chart -->
-        {% if attr.analysis.barchart %}
-        <h3>Bar Chart</h3>
-        <img src="{{ attr.analysis.barchart }}" alt="Bar Chart of {{ attr.attribute }}">
-        {% endif %}
-        {% endif %}
-
-        <!-- Relationship with Target Variable -->
-        <h3>Relationship with Target Variable ({{ target_variable }})</h3>
-        {% if attr.relationship %}
-        {% if attr.relationship.scatter_plot %}
-        <h4>Scatter Plot</h4>
-        <img src="{{ attr.relationship.scatter_plot }}"
-            alt="Scatter Plot of {{ attr.attribute }} vs {{ target_variable }}">
-        {% endif %}
-        {% if attr.relationship.boxplot %}
-        <h4>Boxplot</h4>
-        <img src="{{ attr.relationship.boxplot }}" alt="Boxplot of {{ attr.attribute }} by {{ target_variable }}">
-        {% endif %}
-        {% if attr.relationship.plot %}
-        <h4>Visualization</h4>
-        <img src="{{ attr.relationship.plot }}" alt="Plot of {{ attr.attribute }} vs {{ target_variable }}">
-        {% endif %}
-        {% if attr.relationship.desc_stats %}
-        <h4>Descriptive Statistics by {{ target_variable }}</h4>
-        {{ attr.relationship.desc_stats.to_html(classes='table table-striped') | safe }}
-        {% endif %}
-        {% if attr.relationship.contingency_table %}
-        <h4>Contingency Table</h4>
-        {{ attr.relationship.contingency_table.to_html(classes='table table-striped') | safe }}
-        {% endif %}
-        {% if attr.relationship.pearson %}
-        <h4>Pearson Correlation</h4>
-        <p><strong>Correlation Coefficient:</strong> {{ attr.relationship.pearson.correlation | round(3) }}</p>
-        <p><strong>p-value:</strong> {{ attr.relationship.pearson.p_value | scientific_notation(3) }}</p>
-        {% endif %}
-        {% if attr.relationship.spearman %}
-        <h4>Spearman Correlation</h4>
-        <p><strong>Correlation Coefficient:</strong> {{ attr.relationship.spearman.correlation | round(3) }}</p>
-        <p><strong>p-value:</strong> {{ attr.relationship.spearman.p_value | scientific_notation(3) }}</p>
-        {% endif %}
-        {% if attr.relationship.test %}
-        <h4>{{ attr.relationship.test.name }}</h4>
-        <p><strong>Statistic:</strong> {{ attr.relationship.test.statistic | round(4) }}</p>
-        <p><strong>p-value:</strong> {{ attr.relationship.test.p_value | scientific_notation(4) }}</p>
-        {% endif %}
-        {% if attr.relationship.effect_size %}
-        <h4>Effect Size</h4>
-        {% for key, value in attr.relationship.effect_size.items() %}
-        <p><strong>{{ key.replace('_', ' ').title() }}:</strong> {{ value | round(3) }}</p>
-        {% endfor %}
-        {% endif %}
-        {% if attr.relationship.chi_squared %}
-        <h4>Chi-Squared Test</h4>
-        <p><strong>Chi-Squared Statistic:</strong> {{ attr.relationship.chi_squared.chi2_statistic | round(4) }}</p>
-        <p><strong>p-value:</strong> {{ attr.relationship.chi_squared.p_value | scientific_notation(4) }}</p>
-        <p><strong>Degrees of Freedom:</strong> {{ attr.relationship.chi_squared.degrees_of_freedom }}</p>
-        {% endif %}
-        {% if attr.relationship.cramers_v %}
-        <h4>Cramér's V</h4>
-        <p><strong>Cramér's V:</strong> {{ attr.relationship.cramers_v | round(3) }}</p>
-        <p><strong>Association Strength:</strong> {{ attr.relationship.association_strength }}</p>
-        {% endif %}
-        {% else %}
-        <p>No analysis could be performed for this variable pair.</p>
-        {% endif %}
-    </div>
-    {% endfor %}
-
-</body>
-
-</html>
